+++ app/controllers/auth/callback.js
@@ -1,5 +1,5 @@
 import Controller from '@ember/controller';
 
-export default class CallbackController extends Controller {
+export default class AuthCallbackController extends Controller {
   queryParams = ['code'];
 }
