+++ app/router.js
@@ -10,8 +10,9 @@ export default class Router extends EmberRouter {
 Router.map(function() {
   this.route('mock-login');
   this.route('login', { path: '/aanmelden', });
-  this.route('authorization', { path: '/authorization/callback' });
-  this.route('callback', { path: '/torii/redirect.html' });
+  this.route('auth', { path: '/authorization' }, function () {
+    this.route('callback');
+  });
   this.route('accountless-users', { path: '/onbevoegde-gebruiker', });
 
   // Redirect routes
